requirements.txt
streamlit>=1.25.0
pandas>=1.5.0
pyodbc>=4.0.34
sqlalchemy>=1.4.0
python-dotenv>=0.19.0
azure-sql-database>=1.0.0
azure-identity>=1.13.0